+++ assets/src/apps/delivery/layouts/deck/DeckLayoutView.tsx
@@ -495,6 +495,7 @@ const DeckLayoutView: React.FC<LayoutProps> = ({ pageTitle, pageContent, preview
 
   return (
     <div ref={fieldRef} className={activityClasses.join(' ')}>
+      <style>{`style { display: none !important; }`}</style>
       <DeckLayoutHeader
         pageName={pageTitle}
         userName={currentUserName}