+++ CHANGELOG.md
@@ -8,6 +8,13 @@
 
 ### Enhancements
 
+## 0.13.1 (2021-09-14)
+
+### Bug Fixes
+
+- Fix an issue where platform roles were failing to update on LTI launch
+- Fix an issue that prevents projects page from loading when a project has no collaborators
+
 ## 0.13.0 (2021-09-07)
 
 ### Bug Fixes